Move utils.py into sub-package
